+++ analyzeme/src/testing_common.rs
@@ -1,5 +1,4 @@
-use crate::event_payload::EventPayload;
-use crate::{Event, ProfilingData, Timestamp};
+use crate::{Event, EventPayload, ProfilingData, Timestamp};
 use measureme::{EventId, EventIdBuilder, Profiler, StringId};
 use rustc_hash::FxHashMap;
 use std::borrow::Cow;
@@ -120,13 +119,13 @@ fn process_profiling_data(filestem: &Path, expected_events: &[Event<'static>]) {
 
     // Check iterating forward over the events
     check_profiling_data(
-        &mut profiling_data.iter().map(|e| e.to_event()),
+        &mut profiling_data.iter_full(),
         &mut expected_events.iter().cloned(),
         expected_events.len(),
     );
     // Check iterating backwards over the events
     check_profiling_data(
-        &mut profiling_data.iter().rev().map(|e| e.to_event()),
+        &mut profiling_data.iter_full().rev(),
         &mut expected_events.iter().rev().cloned(),
         expected_events.len(),
     );