+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [10.0.0] - 2021-09-xx
+
+### Changed
+- `analyzeme`: Version-specific parts split out into `decodeme` crate. ([GH-181])
+- `analyzeme`: The crate now supports load both v7 and v8 of the file format. ([GH-181])
+
+## [9.2.0] - 2021-09-13
+### Changed
+- `analyzeme`: Makes a couple of methods in ProfilingData public. ([GH-180])
+
 ## [9.1.2] - 2021-05-21
 ### Added
 - `measureme`: Allow recording interval events without using the drop guard ([GH-159])
@@ -92,6 +102,7 @@
 
 ## [0.2.0] - 2019-04-10
 
+[9.2.0]: https://github.com/rust-lang/measureme/releases/tag/9.2.0
 [9.1.2]: https://github.com/rust-lang/measureme/releases/tag/9.1.2
 [9.1.1]: https://github.com/rust-lang/measureme/releases/tag/9.1.1
 [9.1.0]: https://github.com/rust-lang/measureme/releases/tag/9.1.0
@@ -143,3 +154,5 @@
 [GH-155]: https://github.com/rust-lang/measureme/pull/155
 [GH-156]: https://github.com/rust-lang/measureme/pull/156
 [GH-159]: https://github.com/rust-lang/measureme/pull/159
+[GH-180]: https://github.com/rust-lang/measureme/pull/180
+[GH-181]: https://github.com/rust-lang/measureme/pull/181